Adding to Cart

+++ src/source/services/FileHandler.java
@@ -3,8 +3,6 @@ import java.io.*;
 import java.util.Scanner;
 
 public class FileHandler {
-
-
     public static boolean FileExists(String path){
         File file = new File(path);
         return file.exists();
@@ -24,6 +22,7 @@ public class FileHandler {
 
     public static boolean ClearFile(String path) {
         try {
+            if (!FileExists(path)) return true;
             FileWriter fw = new FileWriter(path);
             fw.close();
             return true;
@@ -35,6 +34,7 @@ public class FileHandler {
 
     public static boolean Remove(String path, String data) {
         try {
+            if (!FileExists(path)) return true;
             Scanner fs = new Scanner(new File(path));
             while(fs.hasNextLine()) {
                 String line = fs.nextLine();
@@ -45,8 +45,7 @@ public class FileHandler {
             fs = new Scanner(tempFile);
             ClearFile(path);
             while (fs.hasNextLine()) {
-                String line = fs.nextLine();
-                Add(path, line);
+                Add(path, fs.nextLine());
             }
             fs.close();
             tempFile.delete();
@@ -57,8 +56,63 @@ public class FileHandler {
         }
     }
 
+    public static boolean DeleteFile(String path) {
+        File file = new File(path);
+        return file.delete();
+    }
 
+    public static boolean CreateDIR(String path) {
+        File dir = new File(path);
+        return dir.mkdirs();
+    }
 
+    public static String GetDataByRef(String path, String ref) {
+        try {
+            Scanner fs = new Scanner(new File(path));
+            while (fs.hasNextLine()) {
+                String line = fs.nextLine();
+                if (line.split("|")[0].equals(ref)) return line;
+            }
+            fs.close();
+            return null;
+        } catch (FileNotFoundException e) {
+            System.out.println(e.getMessage());
+            return null;
+        }
+    }
 
+    public static boolean ModifyData(String path, String oldData, String newData) {
+        try {
+            String tempPath = path.replace(".txt", "_temp.txt");
+            Scanner fs = new Scanner(new File(path));
+            while (fs.hasNextLine()) {
+                String line = fs.nextLine();
+                if (line.equals(oldData)) Add(tempPath, newData);
+                else Add(tempPath, line);
+            }
+            fs.close();
+            File tempFile = new File(tempPath);
+            fs = new Scanner(tempFile);
+            while(fs.hasNextLine()) {
+                Add(path, fs.nextLine());
+            }
+            fs.close();
+            tempFile.delete();
+            return true;
+        } catch (FileNotFoundException e) {
+            System.out.println(e.getMessage());
+            return false;
+        }
+    }
 
+    public static String GenerateUID(int index) {
+        try {
+            Scanner fs = new Scanner(new File("./data/currentUID.txt"));
+            String line = fs.nextLine();
+            return String.valueOf(Integer.parseInt(line.split("|")[index]) + 1);
+        } catch (FileNotFoundException e) {
+            System.out.println(e.getMessage());
+            return null;
+        }
+    }
 }
